feat: Complete Equipment List Expansion system with fighter card integration
Implements issue #693 - Equipment List Expansion system that allows fighters
to access additional equipment based on gang attributes, houses, and fighter
categories through polymorphic rules.

Key features:
- Polymorphic rule system with AND logic for combining multiple conditions
  - Attribute-based rules (e.g., Underdog status)
  - House-specific rules (e.g., House Orlock)
  - Fighter category rules (e.g., Crew fighters)
- Equipment list expansions with cost overrides (different costs for different contexts)
- Integration with equipment editing view showing expansion items
- Fighter card display support for expansion categories
- Detection of actual assigned gear from restricted categories

Technical implementation:
- New models in gyrinx/content/models_/expansion.py to avoid circular imports
- Enhanced ContentEquipmentQuerySet.with_expansion_cost_for_fighter method
- Updated edit_list_fighter_equipment view to include expansion equipment
- Extended has_house_additional_gear and house_additional_gearline_display
  to support three sources: house restrictions, actual assignments, and expansions
- Comprehensive test suite with 16 tests covering all scenarios

This allows game creators to define flexible equipment access rules that
dynamically expand what equipment is available to fighters based on
contextual conditions, making the game system more adaptable.
